Refactor into a plugin JAR and a tools JAR
As we try to leverage the worker API more, we will benefit from some
classes not being on Gradle's classpath, but being available for use in
a worker context. Providing two JARs allows us to keep this separation
more clear.
