+++ modules/gradle-clojure-plugin/src/compatTest/clojure/gradle_clojure/compat_test/test_kit.clj
