+++ modules/gradle-clojure-plugin/src/compatTest/projects/BasicClojureProjectTest/build.gradle
