+++ modules/gradle-clojure-plugin/src/compatTest/projects/BasicClojureProjectTest/src/main/clojure/basic_project/core.clj
