+++ modules/gradle-clojure-plugin/src/compatTest/projects/BasicClojureProjectTest/src/test/clojure/basic_project/core_test.clj
