+++ modules/gradle-clojure-plugin/src/compatTest/projects/IncrementalCompilationTest/build.gradle
