+++ modules/gradle-clojure-plugin/src/compatTest/projects/JavaAndClojureInOneSourceSetTest/build.gradle
