+++ modules/gradle-clojure-plugin/src/compatTest/projects/JavaAndClojureInOneSourceSetTest/src/main/clojure/clj_example/core.clj
