+++ modules/gradle-clojure-plugin/src/compatTest/projects/JavaAndClojureInOneSourceSetTest/src/main/java/javaExample/Example.java
