+++ modules/gradle-clojure-plugin/src/compatTest/projects/MixedClojureJavaTest/build.gradle
