+++ modules/gradle-clojure-plugin/src/compatTest/projects/MixedClojureJavaTest/src/cljSS/clojure/cljSS/Example.clj
