+++ modules/gradle-clojure-plugin/src/compatTest/projects/MixedClojureJavaTest/src/javaSS/java/javaSS/Test.java
