+++ modules/gradle-clojure-plugin/src/compatTest/projects/MixedJavaClojureTest/build.gradle
