+++ modules/gradle-clojure-plugin/src/compatTest/projects/MixedJavaClojureTest/src/cljSS/clojure/cljSS/core.clj
