+++ modules/gradle-clojure-plugin/src/compatTest/projects/MixedJavaClojureTest/src/javaSS/java/javaSS/Example1.java
