+++ modules/gradle-clojure-plugin/src/compatTest/projects/MultipleSourceSetsTest/build.gradle
