+++ modules/gradle-clojure-plugin/src/compatTest/projects/MultipleSourceSetsTest/src/ss1/clojure/ss1/core.clj
