+++ modules/gradle-clojure-plugin/src/compatTest/projects/MultipleSourceSetsTest/src/ss2/clojure/ss2/core.clj
