+++ modules/gradle-clojure-plugin/src/compatTest/projects/MultipleSourceSetsTest/src/test/clojure/test/core_test.clj
