+++ modules/gradle-clojure-plugin/src/compatTest/projects/TestFailureFailsBuildTest/build.gradle
