+++ modules/gradle-clojure-plugin/src/compatTest/projects/TestFailureFailsBuildTest/src/main/clojure/basic_project/core.clj
