+++ modules/gradle-clojure-plugin/src/compatTest/projects/TestFailureFailsBuildTest/src/test/clojure/basic_project/core_test.clj
