+++ modules/gradle-clojure-plugin/src/main/java/gradle_clojure/plugin/internal/ClojureEval.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package gradle_clojure.clojure.tasks.internal;
+package gradle_clojure.plugin.internal;
 
 import java.io.File;
 import java.net.MalformedURLException;
