+++ modules/gradle-clojure-plugin/src/main/java/gradle_clojure/plugin/tasks/ClojureSourceSet.java
@@ -13,9 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package gradle_clojure.clojure.tasks;
+package gradle_clojure.plugin.tasks;
 
-import org.gradle.api.tasks.compile.BaseForkOptions;
+import org.gradle.api.Action;
+import org.gradle.api.file.SourceDirectorySet;
 
-public class ClojureForkOptions extends BaseForkOptions {
+public interface ClojureSourceSet {
+  SourceDirectorySet getClojure();
+
+  ClojureSourceSet clojure(Action<? super SourceDirectorySet> configureAction);
 }
