+++ modules/gradle-clojure-tools/build.gradle
@@ -1,0 +1,18 @@
+plugins {
+  id 'java'
+  id 'gradle-clojure.clojure' version '0.1.0'
+  id 'maven-publish'
+}
+
+dependencies {
+  // gradle
+  compileOnly gradleApi()
+}
+
+publishing {
+  publications {
+    main(MavenPublication) {
+      from components.java
+    }
+  }
+}
