+++ modules/gradle-clojure-tools/src/main/java/gradle_clojure/tools/internal/LineProcessingOutputStream.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package gradle_clojure.clojure.tasks.internal;
+package gradle_clojure.tools.internal;
 
 import java.io.OutputStream;
 import java.nio.ByteBuffer;
