+++ settings.gradle
@@ -1,1 +1,3 @@
 rootProject.name = 'gradle-clojure'
+include 'modules:gradle-clojure-plugin'
+include 'modules:gradle-clojure-tools'
